ml/NER_tesseractV3/requirements.txt
transformers==4.41.1
torch>=1.12.0
datasets
seqeval
scikit-learn
matplotlib
pandas
paddleocr
fastapi
uvicorn
Pillow
pyyaml
